refactor: move NavButtons to EmployeesPrivatePage
