--- src/components/UI/EmployeesNavButtons/index.ts
@@ -1,1 +1,0 @@
-export { EmployeesNavButtons } from './EmployeesNavButtons';
