+++ src/pages/EmployeesPrivatePage/EmployeesPrivatePage.tsx
@@ -1,9 +1,9 @@
 import React, { useEffect } from 'react';
 import { Outlet, useLocation, useNavigate } from 'react-router-dom';
-import { EmployeesNavButtons } from '../../components/UI/EmployeesNavButtons';
 import { RoutePath } from '../../constants/routeVariables';
 import { checkIdRoute } from '../../utils/checkIdRoute';
 import * as Styled from './EmployeesPrivatePage.styles';
+import { NavButtons } from './NavButtons/NavButtons';
 
 const EmployeesPrivatePage = () => {
   const location = useLocation();
@@ -18,7 +18,7 @@ const EmployeesPrivatePage = () => {
   return (
     <main>
       <Styled.ContainerWrapper maxWidth="xl">
-        <EmployeesNavButtons />
+        <NavButtons />
         <Outlet />
       </Styled.ContainerWrapper>
     </main>
