+++ src/pages/EmployeesPrivatePage/NavButtons/NavButtons.styles.ts
